[New Historator library added for archiving script running session values]
New Features:
New Archive functionality for holding parameter information used at run time for later retrivial. This uses the Script Executors method name and the declaring type to provide a name to associate the parameters with. Allowing for the same data to be used to retrieve it at a later time.

These Files can be Added to the historator.

Functionality Changes:
* The CLI UI will now append the value of a previous running sessions string value to the end of its name inside of paranthesis.
* The Appended value will be used if no input is provided.
* The previous session data is retrieve from archive and is saved out to the archive once all parameters have been retrieved (regardless of any entries changing).